Remove obsolete code
Also, mild refactors + upgrade rust toolchain

Fixes #118
