+++ .buildkite/docker/rust/Dockerfile
@@ -1,25 +1,46 @@
-FROM rust:slim
+FROM debian:buster-slim
 
-# Install additional packages
-RUN apt-get update \
-    && apt-get install -y --no-install-recommends \
-        clang \
+# System packages
+RUN set -eux; \
+    apt-get update; \
+    apt-get install -y --no-install-recommends \
+        ca-certificates \
         curl \
         file \
-        libclang-dev \
-        libgmp-dev \
+        gcc \
+        git \
+        libc6-dev \
+        liblzma-dev \
         libssl-dev \
-        llvm-dev \
-        m4 \
         make \
         pkg-config \
-    && apt-get autoremove \
-    && rm -rf /var/lib/apt/lists/*
+        ; \
+    apt-get autoremove; \
+    rm -rf /var/lib/apt/lists/*
 
-RUN rustup component add rustfmt \
-    && cargo install sccache cargo-deb \
-    && rm -rf /usr/local/cargo/registry \
-    && rm /usr/local/cargo/.package-cache
+# Rust toolchain
+# Make sure this is in sync with rust-toolchain!
+ENV RUST_VERSION=nightly-2020-05-11 \
+    CARGO_HOME=/usr/local/cargo \
+    PATH=/usr/local/cargo/bin:$PATH \
+    RUSTUP_HOME=/usr/local/rustup \
+    RUSTUP_VERSION=1.21.1 \
+    RUSTUP_SHA256=ad1f8b5199b3b9e231472ed7aa08d2e5d1d539198a15c5b1e53c746aad81d27b
+
+RUN set -eux; \
+    curl -LOf "https://static.rust-lang.org/rustup/archive/${RUSTUP_VERSION}/x86_64-unknown-linux-gnu/rustup-init"; \
+    echo "${RUSTUP_SHA256} *rustup-init" | sha256sum -c -; \
+    chmod +x rustup-init; \
+    ./rustup-init -y --no-modify-path --profile minimal --default-toolchain $RUST_VERSION; \
+    rm rustup-init; \
+    chmod -R a+w $RUSTUP_HOME $CARGO_HOME; \
+    rustup --version; \
+    cargo --version; \
+    rustc --version; \
+    rustup component add clippy rustfmt; \
+    cargo install sccache cargo-deny; \
+    rm -rf /usr/local/cargo/registry; \
+    rm /usr/local/cargo/.package-cache;
 
 VOLUME /cache
 ENV SCCACHE_DIR=/cache/sccache \
