+++ .buildkite/pipeline.yaml
@@ -7,8 +7,8 @@ steps:
       production: "true"
       platform: "linux"
     env: &build-docker
-      DOCKER_IMAGE: "gcr.io/opensourcecoin/radicle-link-seedling-build@sha256:f53c35f65541d91b7665003a6362be48d6b2687a5a9e46adaadccaf3129da51f"
-      DOCKER_FILE: .buildkite/docker/rust-nightly/Dockerfile
+      DOCKER_IMAGE: "gcr.io/opensourcecoin/radicle-link-seedling-build@sha256:f8280c841387ea2ced917366b78af9870e3330ca205478a998943f3c8f3c948f"
+      DOCKER_FILE: .buildkite/docker/rust/Dockerfile
 
   - label: "Lint and clip"
     commands:
