+++ librad/Cargo.toml
@@ -11,36 +11,20 @@ bit-vec = "0.6"
 bs58 = "0.3"
 bytes = "0.5"
 directories = "2.0"
-failure = { version = "0.1", features = ["backtrace"] }
-fnv = "1.0"
-futures = { version = "0.3", features = ["thread-pool"] }
-futures_codec = { version = "0.4", features = ["cbor"] }
 futures-timer = "3.0"
-git2 = { version = "0.12", default-features = false, features = [] }
 governor = "0.2"
 hex = "0.4"
 lazy_static = "1.4"
-# Need to mention libgit2-sys explicitly in order to disable openssl
-# (incompatible license)
-libgit2-sys = { version = "0.11", default-features = false, features = [] }
 log = "0.4"
 multibase = "0.8"
 multihash = "0.10"
-nettle = { version = "5.0", features = ["vendored"]  }
 nonempty = "0.2"
 olpc-cjson = "0.1"
 percent-encoding = "2.1"
-quinn = { version = "0.6", default-features = false, features = ["tls-rustls"] }
-radicle-keystore = { path = "../radicle-keystore" }
-radicle-surf = { git = "https://github.com/radicle-dev/radicle-surf", rev = "33d4980422bc6a0cb48ccdc6d4b893f9e45bc2b9" }
 rand = "0.7"
 rand_pcg = "0.2"
 rcgen = "0.7"
 regex = "1.3"
-rustls = { version = "0.17", features = ["logging", "dangerous_configuration"] }
-secstr = "0.3"
-sequoia-openpgp = "0.12"
-serde = { version = "1.0", features = ["derive"] }
 serde_cbor = "0.10"
 serde_json = "1.0"
 serde_repr = "0.1"
@@ -48,18 +32,70 @@ sha2 = "0.8"
 sodiumoxide = "0.2"
 tempfile = "3.1"
 thiserror = "1.0"
-tokio = { version = "0.2", features = ["full"] }
-tokio-util = { version = "0.3.0", features = ["compat"] }
+secstr = "0.3"
 tracing = "0.1"
 tracing-futures = "0.2"
-url = { version = "2.1", features = ["serde"] }
 urltemplate = "0.1"
 webpki = "0.21"
-yasna = { version = "0.3", features = ["bit-vec"] }
+
+[dependencies.futures]
+version = "0.3"
+features = ["thread-pool"]
+
+[dependencies.futures_codec]
+version = "0.4"
+features = ["cbor"]
+
+[dependencies.git2]
+version = "0.12"
+default-features = false
+features = []
+
+[dependencies.libgit2-sys]
+version = "0.11"
+default-features = false
+features = []
+
+[dependencies.quinn]
+version = "0.6"
+default-features = false
+features = ["tls-rustls"]
+
+[dependencies.radicle-keystore]
+path = "../radicle-keystore"
+
+[dependencies.radicle-surf]
+git = "https://github.com/radicle-dev/radicle-surf"
+rev = "33d4980422bc6a0cb48ccdc6d4b893f9e45bc2b9"
+
+[dependencies.rustls]
+version = "0.17"
+features = ["logging", "dangerous_configuration"]
+
+[dependencies.serde]
+version = "1.0"
+features = ["derive"]
+
+[dependencies.tokio]
+version = "0.2"
+features = ["full"]
+
+[dependencies.tokio-util]
+version = "0.3"
+features = ["compat"]
+
+[dependencies.url]
+version = "2.1"
+features = ["serde"]
+
+[dependencies.yasna]
+version = "0.3"
+features = ["bit-vec"]
 
 [dev-dependencies]
 anyhow = "1.0"
 env_logger = "0.7"
+fnv = "1.0"
 futures-await-test = "0.3"
 futures_ringbuf = "0.2"
 matches = "0.1"
