+++ librad/src/lib.rs
@@ -19,7 +19,6 @@
 #![feature(bool_to_option)]
 
 extern crate radicle_keystore as keystore;
-extern crate sequoia_openpgp as pgp;
 extern crate sodiumoxide;
 #[macro_use]
 extern crate lazy_static;
@@ -28,17 +27,13 @@ pub use radicle_surf as surf;
 
 pub mod git;
 pub mod hash;
+pub mod internal;
 pub mod keys;
 pub mod meta;
 pub mod net;
 pub mod paths;
 pub mod peer;
-pub mod project;
-pub mod sync;
 pub mod uri;
-
-pub(crate) mod canonical;
-pub(crate) mod channel;
 
 pub fn init() -> bool {
     sodiumoxide::init().is_ok()
