+++ librad/src/meta/serde_helpers.rs
@@ -79,45 +79,3 @@ pub(crate) mod urltemplate {
         Ok(v.map(|Wrapper(a)| a))
     }
 }
-
-pub(crate) mod pgp_fingerprint {
-    use pgp::Fingerprint;
-    use serde::{Deserialize, Deserializer, Serializer};
-
-    pub fn serialize<S>(fingerprint: &Fingerprint, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        serializer.serialize_str(&fingerprint.to_hex())
-    }
-
-    pub fn serialize_opt<S>(opt: &Option<Fingerprint>, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        if let Some(fingerprint) = opt {
-            serialize(fingerprint, serializer)
-        } else {
-            serializer.serialize_none()
-        }
-    }
-
-    pub fn deserialize<'de, D>(deserializer: D) -> Result<Fingerprint, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let string = String::deserialize(deserializer)?;
-        Fingerprint::from_hex(&string).map_err(serde::de::Error::custom)
-    }
-
-    pub fn deserialize_opt<'de, D>(deserializer: D) -> Result<Option<Fingerprint>, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        #[derive(Deserialize)]
-        struct Wrapper(#[serde(deserialize_with = "deserialize")] Fingerprint);
-
-        let v = Option::deserialize(deserializer)?;
-        Ok(v.map(|Wrapper(a)| a))
-    }
-}
