--- rad2/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "rad2"
-version = "0.1.0"
-authors = ["The Radicle Team <dev@radicle.xyz>"]
-edition = "2018"
-description = "The Official Unoffical Radicle CLI"
-license = "GPL-3.0-or-later"
-
-[dependencies]
-anyhow = "1.0"
-git2 = { version = "0.12", default-features = false, features = [] }
-# Need to mention libgit2-sys explicitly in order to disable openssl
-libgit2-sys = { version = "0.11", default-features = false, features = [] }
-librad = { path = "../librad" }
-radicle-keystore = { path = "../radicle-keystore" }
-secstr = "0.3"
-serde = "1.0"
-serde_yaml = "0.8"
-structopt = "0.3"
-tempfile = "3.1"
-thiserror = "1.0"
-url = "2.1"
-
