--- rad2/README.md
@@ -1,1 +1,0 @@
-**rad2** is the official unofficial CLI for `radicle-link`
