+++ rust-toolchain
@@ -1,1 +1,1 @@
-nightly-2020-01-16
+nightly-2020-05-11
